requirements.txt
streamlit
yfinance
vaderSentiment
newsapi-python
pandas
matplotlib
ta
scikit-learn
transformers
torch
plotly